Make NOR reading code good

+++ PS5 NOR Modifier/Offsets.cs
@@ -2,11 +2,11 @@ namespace PS5_NOR_Modifier;
 
 public static class Offsets
 {
-    public const long One = 0x1c7010; // TODO: These two need better names
-    public const long Two = 0x1c7030;
-    public const long WiFiMAC = 0x1C73C0;
-    public const long LANMAC = 0x1C4020;
-    public const long Serial = 0x1c7210;
-    public const long Variant = 0x1c7226;
-    public const long MoboSerial = 0x1C7200;
+    public const int One = 0x1c7010; // TODO: These two need better names
+    public const int Two = 0x1c7030;
+    public const int WiFiMAC = 0x1C73C0;
+    public const int LANMAC = 0x1C4020;
+    public const int Serial = 0x1c7210;
+    public const int Variant = 0x1c7226;
+    public const int MoboSerial = 0x1C7200;
 }
